// src/types/enums.ts
//로컬스토리지 테마
export enum LOCAL_THEME {
  LOCAL_KEY = 'theme',
  LIGHT_VALUE = 'light',
  DARK_VALUE = 'dark',
}

//Header, Sidebar Menu 네비게이션 툴팁
export enum TOOLTIP {
  MAIL = 'mail',
  THEME = 'theme',
  RESET = 'reset',
}

//Sidebar Menu open/close 상태
export enum SIDEBAR_STATUS {
  OPEN = 'open',
  CLOSE = 'close',
}